add handleEditPost and inhancment project

+++ viewPost.php
@@ -32,28 +32,50 @@ if (mysqli_num_rows($result) == 1) {
   <div class="container">
     <div class="row">
       <div class="col-md-12">
+        <?php
+        if (isset($_SESSION['errors'])) {
+          foreach ($_SESSION['errors'] as $error) {
+            echo "<div class='alert alert-danger'>$error</div>";
+          }
+          unset($_SESSION['errors']);
+        }
+        if (isset($_SESSION['success'])) {
+          foreach ($_SESSION['success'] as $success) {
+            echo "<div class='alert alert-success'>$success</div>";
+          }
+          unset($_SESSION['success']);
+        }
+        ?>
+
         <div class="section-heading">
-          <h2><?php echo $post['title']; ?></h2>
+          <h2>Post</h2>
         </div>
       </div>
-      <div class="col-md-6">
-        <div class="right-image">
-          <img src="assets/images/postimage/<?php echo $post['image']; ?>" alt="">
-        </div>
-      </div>
-      <div class="col-md-6">
-        <div class="left-content">
-          <h4><?php echo $post['title']; ?></h4>
-          <p class='post-view'><?php echo $post['body']; ?></p>
-          <p><?php echo $post['created_at']; ?></p>
-
-          <div class="d-flex justify-content-center">
-            <a href="editPost.php?id=<?php echo $post['id']; ?>" class="btn btn-success mr-3 "> edit post</a>
-
-            <a href="handel/handeldeletePost.php?id=<?php echo $post['id']; ?>" class="btn btn-danger "> delete post</a>
+      <?php
+      if (!empty($post)) {
+      ?>
+        <div class="col-md-6">
+          <div class="right-image">
+            <img src="assets/images/postimage/<?php echo $post['image']; ?>" alt="">
           </div>
         </div>
-      </div>
+        <div class="col-md-6">
+          <div class="left-content">
+            <h4><?php echo $post['title']; ?></h4>
+            <p class='post-view'><?php echo $post['body']; ?></p>
+            <p><?php echo $post['created_at']; ?></p>
+
+            <div class="d-flex justify-content-center">
+              <a href="editPost.php?id=<?php echo $post['id']; ?>" class="btn btn-success mr-3 "> edit post</a>
+
+              <a href="handel/handeldeletePost.php?id=<?php echo $post['id']; ?>" class="btn btn-danger "> delete post</a>
+            </div>
+          </div>
+        </div>
+      <?php
+      } else {
+        echo "<div class='alert alert-danger w-50'>$msg</div>";
+      } ?>
     </div>
   </div>
 </div>
